Refactor X button out to PromoModal.jsx. Move Insights and Plus PromoModalContent out to shared-components
