+++ app/shared-components/ModalContent/PlusPromoModalContent/index.js
@@ -1,0 +1,16 @@
+/**
+ * Point of entry index.js file for PlusPromoModalContent Component
+ *
+ * Ghostery Browser Extension
+ * https://www.ghostery.com/
+ *
+ * Copyright 2019 Ghostery, Inc. All rights reserved.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0
+ */
+
+import PlusPromoModalContent from './PlusPromoModalContent';
+
+export default PlusPromoModalContent;
