+++ app/shared-components/ModalContent/PremiumPromoModalContent/PremiumPromoModalContent.scss
