+++ manifest.json
@@ -105,5 +105,4 @@
 		"cliqz/offers-reminder/index.html",
 		"cliqz/popup-notification/images/*"
 	],
-	"debug": true
 }
